declutting and updating unit test for netcat to run individually.
