+++ makefile
@@ -1,4 +1,4 @@
-.PHONY: help clean build test format tag version checkGitClean mermaid bump-patch bump-minor bump-major
+.PHONY: help clean build test test-netcat-client-only test-netcat-server-only format tag version checkGitClean mermaid bump-patch bump-minor bump-major release
 
 .DEFAULT_GOAL := help
 
@@ -19,11 +19,11 @@ build:  ## Build the project in release mode
 test:  ## Run tests (excluding netcat tests)
 	swift test
 
-test-netcat:  ## Run all tests including netcat integration tests
-	RUN_NETCAT_TESTS=1 swift test
+test-netcat-client-only:  ## Run only netcat client integration tests
+	RUN_NETCAT_CLIENT_TESTS=1 swift test --filter "connectClientToNetCat"
 
-test-netcat-only:  ## Run only netcat integration tests
-	RUN_NETCAT_TESTS=1 swift test --filter "NetCat"
+test-netcat-server-only:  ## Run only netcat server integration tests
+	RUN_NETCAT_SERVER_TESTS=1 swift test --filter "connectServerToNetCat"
 
 format:  ## Format code using swift-format with explicit config
 	swift-format --configuration $(CONFIG) format --in-place --recursive spm/Sources
